pyproject.toml
[project]
name = "curegemma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "openai>=1.101.0",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]
